# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HealthCaring365 Demo API",
    "version": "1.0.0",
    "description": "Demo healthcare API with rich CRUD operations for Postman practice."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "PatientSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "mrn": { "type": "string" },
          "name": { "type": "string" },
          "ageYears": { "type": "integer" },
          "gender": { "type": "string" },
          "bloodGroup": { "type": "string" },
          "primaryCondition": { "type": "string", "nullable": true },
          "lastEncounter": {
            "type": "object",
            "nullable": true,
            "properties": {
              "encounterId": { "type": "string" },
              "department": { "type": "string" },
              "date": { "type": "string", "format": "date-time" },
              "status": { "type": "string" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login and get API key",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "username": { "type": "string" },
                        "displayName": { "type": "string" }
                      }
                    },
                    "apiKey": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "/auth/logout": {
        "post": {
          "summary": "Logout user",
          "description": "Client should delete its API key after calling this endpoint.",
          "responses": {
            "200": {
              "description": "Logout successful"
            }
          }
        }
      }

      
    },
...
    "/api/patients": {
      "get": {
        "summary": "List patient summaries",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Search by name / id / MRN"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of patient summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PatientSummary" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new patient",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatientBundle" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientBundle" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Patient already exists" }
        }
      }
    },
    "/api/patients/paged": {
      "get": {
        "summary": "Paged patient summaries",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" },
                    "total": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/PatientSummary" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/patients/stats": {
      "get": {
        "summary": "Patient statistics",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Stats by gender, blood group, condition"
          }
        }
      }
    },
    "/api/patients/{id}/full": {
      "get": {
        "summary": "Get full patient record",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Full patient bundle",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientBundle" }
              }
            }
          },
          "404": { "description": "Patient not found" }
        }
      }
    },
    "/api/patients/{id}/encounters": {
      "get": {
        "summary": "Get encounters for patient",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of encounters" },
          "404": { "description": "Patient not found" }
        }
      }
    },
    "/api/patients/{id}/vitals-timeline": {
      "get": {
        "summary": "Get vitals timeline for patient",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of vitals points" },
          "404": { "description": "Patient not found" }
        }
      }
    },
    "/api/patients/{id}": {
      "put": {
        "summary": "Replace patient",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatientBundle" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated bundle" },
          "400": { "description": "Validation error" },
          "404": { "description": "Patient not found" }
        }
      },
      "patch": {
        "summary": "Partial update patient",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Merged bundle" },
          "404": { "description": "Patient not found" }
        }
      },
      "delete": {
        "summary": "Delete patient",
        "tags": ["Patients"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "Patient not found" }
        }
      }
    },
    "/api/appointments": {
      "get": {
        "summary": "List appointments with filters",
        "tags": ["Appointments"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "patientId", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string" } },
          { "name": "dateTo", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Array of appointments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Appointment" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create appointment",
        "tags": ["Appointments"],
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Appointment" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Appointment" }
              }
            }
          },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/appointments/{id}": {
      "get": {
        "summary": "Get appointment by id",
        "tags": ["Appointments"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Appointment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Appointment" }
              }
            }
          },
          "404": { "description": "Appointment not found" }
        }
      },
      "put": {
        "summary": "Replace appointment",
        "tags": ["Appointments"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Appointment" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated appointment" },
          "400": { "description": "Validation error" },
          "404": { "description": "Appointment not found" }
        }
      },
      "delete": {
        "summary": "Delete appointment",
        "tags": ["Appointments"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "Appointment not found" }
        }
      }
    },
    "/api/appointments/{id}/cancel": {
      "patch": {
        "summary": "Cancel appointment",
        "tags": ["Appointments"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Appointment cancelled" },
          "404": { "description": "Appointment not found" }
        }
      }
    },
...
  }
}